+++ lib/dartemis.dart
@@ -11,7 +11,7 @@ part 'src/core/utils/utils.dart';
 part 'src/core/utils/object_pool.dart';
 
 part 'src/core/component.dart';
-part 'src/core/component_mapper.dart';
+part 'src/core/mapper.dart';
 part 'src/core/component_manager.dart';
 part 'src/core/component_type.dart';
 part 'src/core/component_type_manager.dart';